CXO-7741 - Testing plus small change. Providing upgraders.

+++ comu/com.energyict.mdc.device.topology/src/main/java/com/energyict/mdc/device/topology/impl/kpi/RegisteredDevicesKpiScoreImpl.java
@@ -37,7 +37,7 @@ public class RegisteredDevicesKpiScoreImpl implements RegisteredDevicesKpiScore 
     }
 
     @Override
-    public long getTarget(int target) {
+    public long getTarget(long target) {
         return Math.round(getTotal().divide(BigDecimal.valueOf(100)).multiply(BigDecimal.valueOf(target)).doubleValue());
     }
