+++ comu/com.energyict.mdc.device.topology/src/main/java/com/energyict/mdc/device/topology/kpi/RegisteredDevicesKpiScore.java
@@ -33,5 +33,5 @@ public interface RegisteredDevicesKpiScore extends Comparable<RegisteredDevicesK
 
     BigDecimal getRegistered();
 
-    long getTarget(int target);
+    long getTarget(long target);
 }
